Bugfix - NSX-T Firewall creation eve-consistency
This changeset contains the implementation of firewall rules through
REST api instead of SDK.

- Batch creation of Firewall rules starts failing in NSX-T 2.4
In this version of NSX-T the database consistency mechanism was improved,
thus this issue was not experienced in 2.3.
- Another change is making the optionl  parameter _revision mandatory.
- NSX-T Firewall Section Rule Updates
When invoked on a section with a large number of rules,
this API is supported only at low rates of invocation
(not more than 4-5 times per minute). The typical latency of this API with
about 1024 rules is about 4-5 seconds.
This API should not be invoked with large payloads at automation speeds.
More than 50 rules with a large number of rule references is not supported.
Instead, to create sections,
use: POST /api/v1/firewall/sections To create rules,
use: POST /api/v1/firewall/sections/<section-id>/rules