Set timeout for vapi requests_connector
The default timeout being None means endless waiting for a request to
finish. This leads to a thread taking a lock endlessly, which, after
some time, will lead to all worker threads hanging on that lock and the
agent doing nothing.

While request's Session doesn't support a timeout, the vapi does. It
uses it explicitly on every request it makes. One can still use a
different one per request by providing the timeout parameter to the
request.