Add an option to sync security groups but disable them
Disabling := sync NSX-T Groups by skipping membership criteria configuration.

For the migration scenario:
1. Run the Management Agent.
2. Run in parallel the CLI with a security groups sync value.
to set to '*' and CFG option groups_dynamic_members_off=True.
3. Run the Policy Agent with groups_dynamic_members_off=False.
By performing 1-3 the agent will synchronize both the policy and management
API objects to the Neutron revisions. Management objects will remain active
until there is an update on them, which will cause an update on the policy
object that will activate the group membership, increase the revision number
and finally remove the management object as it is going to have lower version.