Unify group memebrship criteria for 2.4.2, 2.5.2 and 3.1.0
A Segment Port is an automatically created policy object for
Logical Port attached to a Segment.

Issue
As Segment Ports got introduced in NSX-T 2.5.2 they cannot
be used as membership criteria in a Policy Group in 2.4.2.

Resolution
In order to keep the membership criteria simplified, I change
the membership criteria to use Logical Port instead of the
Segment Port. This solution will behave in a similar way
across 2.4.2, 2.5.2 and 3.1.0.

Migration
A migration to Segment Port membership criteria can happen
all 2.4.2 environments got an update. The fix will be added
to the agent and will be applied on the first full sync.
The transition will be smooth and there should not be any
side effects or interruption.