Add default logging policy
The default logging policy REJECTs the traffic form
the ports with enabled logging and generate logs for this rejected
traffic.
It is required as the security groups only allow traffic and when
logging is enabled for them and the traffic does not match it is
termiated by the Default L2 Section. Now with the default logging
policy we place it right before the Default L2 Section which enables
the NSX manager to capture the DWF debug logs.

The policy is active only for the ports enabled for logging.
The policy uses Logging Group with memberhsip criteria by tag with
scope := logging, tag := enabled