Remove Ovs-Hybrid-Plug flag
The flag is for custom behaviour specific to
open-v-switch, which we do not want/need to trigger,
so we rather remove it.

Specifically it causes neutron to send a network-vif-plugged event
during live migration after port profile is updated with "migrating-to:"

That in turn causes a the port to be shut-down for a short period,
which we rather want to avoid.