Unpersist MP-TO-POLICY migration before running the brownfield tests
To avoid running the mp-to-policy migration all over again, once a nsxv3 agent get restarted, a tag was introduced indicating
wether the migration was already executed succesfully. As consequence the brownfield tests work only once because on nsxt side a
tag is set indicating the succesfull migration. To overcome this, the migration tag needs to be removed before running a migration