realize port for multiple binding based on active binding from other host
The only way to ensure the port is pre-created correctly is in the get_network_bridge
rpc call from neutron. But for live-migration ports, they don't exist yet for the target host.
In this case, we fetch the (active) port from any host (the source host) and use it's information
to sync the port.

This way, we should be sure that the port exists before migration is happening, therefor
mitigating race condition causing duplicate ports.

Also, if the "migrated" port is found existing on the nsx-t manager but with another
vlan tag, delete it before re-create, instead of "update" it since update never worked
when chaning the underlying segment.