NSXT Bulk API for realizing SGs
For ports with more than 27 SGs, the approach of adding them as static
member one by one results in a rpc-timout. Thus ports rely on the
update_sg_group method executed after a port binding.

With this commit, we use the nsxt bulk API to add the staticmembership
of ports directly after binding the port.