Fetch all information for SG realization at once
With this commit, the number of rpc calls required to fetch all
information from the neutron DB in order to realize the SG Group is reduced from three to
one. Additionally the number of SQL queries reduces by one. It is worth
mentioning that this query was part of in the slow-query-log once in a
while.