Add IPv6 address of metadata service to sg rules
So far only DHCPv6 and ICMPv6 traffic is refelcted in the default
infrastructure rules of NSX-T. The IPv6 address of the metadata
service requires clearing as well.

In order to reflect those changes in NSX-T, it requires a check if
the rule configuration has changed. So far only creation of missing
policies is supported.