[Caracal] Fix Unit Tests
Before loading the NSX-T Operation API default configuration settings
must be loaded as of part of the version migration.

Remove the subnetpool in the DB unit tests as this is not needed for
a subnet creation. Without the subnetpool, the db unit test test_port_sg
fails as the result of a DB query returns the cartesian product between
the tables SecurityGroupPortBindings and StandardAttributes. Thus the
length of the result set is lowered from 9 to 8 as there are only 8
objects with standard attributes.

This is a minor bug which will be fixed in a separate PR.