[Caracal] Update github workflow and Tox
Loosen constraints for the installation of neutron and
neutron-lib in the test-requirements.txt as the upper bound was set to
Neutron Yoga release. Within the tox configuration request the
installation of yoga from the current stable/2024.1-m3 branch.
As Tox version 4 treats '#' as comments a '\' is required.

Change base branch to Caracal release. Furthermore set the base python
version to 3.10.

Update the github-runner to the latest version, as the reason for
sticking to an older version is no longer true.

Loosen constraints for the VIRTUALENV_PIP version. Sticking with pinned
version causes the tox execution to fail.