[bugfix] Fixing JobList that was not handling duplicate jobs correctly
Operations on address_groups will use callbacks that get a dictionary
as parameter, and not just a OpenStackID.

For those jobs, we need to take the content of that dictionary into
account, when we try to minimize the number of identical jobs to run.

JobList had a bug, that when adding a new job and another job
with different parameters in its dictionary was already present, it
was not correctly recognized and the new job was also immediately
allowed to run.

Also adding some tests for that.