fix: Nova fetch realization status
In some cases, Nova does not wait for the network-realized custom event,
as this may be sent too early. Therefore, an API extension was developed
to allow checking the status of a segment port explicitly when the event
is not available.

The check for whether a port is realized or not should match the logic
used in the custom event sent once realization is complete.
Thus, we remove the publish_state.